requirements.txt
Django==3.2.9
Pillow==8.4.0
django-ckeditor==6.1.0
django-mptt==0.13.4
django-admin-thumbnails==0.2.6